refactor: :recycle: move files and urls actions to own files
